Decoupled rectangles from images
This change makes it possible to put multiple font faces in a single texture (to be implemented in a future commit)

Each Face has a std::vector<Image>. To pack multiple std::vector<Image>, we have to somehow merge them into one std::vector. Images are move-only so we would have to move them all out of the Face, pack them, then move the images back into the Face which is really messy. Using a separate std::vector<RectPx> we can simply copy the rectangles. RectPx is TriviallyCopyable so std::memcpy can be used.