Import game assets
A few months ago (using an old version of this game engine), I made a game. I finished the game logic and some of the game view but thats it. I've decided that I'm going to rewrite the game and complete it with this engine. The game is 2D so I should be able to use the UI system in the game view.