Abandon UI::Table
I'm really disappointed. It just wasn't going to work. My UI system is starting to crumble. This is what happened to my 3D rendering system. It got too big, started to fail apart and then I replaced it with Ogre3D. I really don't want to abandon the UI system and replace it with CEGUI because it's just so much code. The UI system is really the only system in this game engine! Everything else is just utilities. I'll try to create the game view without table.

Maybe the problem is that I just can't build big complex systems. I don't thoroughly plan things out and leave room for expansion.

If anyone reading this commit message wants to mentor me, please contact me because that would be amazing!